Extract fuse structs
